Bug 1940142: prepare correct clouds.yaml data
In OCP versions <4.6 we included the original host cacert path in
clouds.yaml. It is wrong because internally we always use
"/etc/kubernetes/static-pod-resources/configmaps/cloud-config/ca-bundle.pem".

The issue was fixed in the installer [1] in 4.6 but we still have to
replace incorrect values in the generated clouds.yaml for the
previous versions.

This commit checks if the value is not correct and fixes it in the root
secret.

[1] openshift/installer#4227

Co-Authored-By: Mike Fedosin <mfedosin@redhat.com>